homepanorama_root/requirements.txt
Django
django-cors-headers
django-mongodb-engine
djangorestframework
django-cors-headers
djangotoolbox
# djongo
gunicorn==20.0.4
jsonschema==3.2.0
oauth2client==4.1.3
oauthlib==3.1.0
pexpect
python-dateutil